Test some LD

+++ src/Core/Emulator/MemoryKeyPoint.cs
@@ -2,17 +2,17 @@ namespace GBMU.Core;
 
 public class MemoryKeyPoint
 {
-	public uint addr { get; set; }
-	public uint size { get; set; }
+	public ushort addr { get; set; }
+	public ushort size { get; set; }
 
-	public MemoryKeyPoint(uint addr, uint size)
+	public MemoryKeyPoint(ushort addr, ushort size)
 	{
 		this.addr = addr;
 		this.size = size;
 	}
 
-	public uint EndAddr => addr + size - 1;
+	public ushort EndAddr => (ushort)(addr + size - 1);
 
-	public bool IsInRange(uint addr) => addr >= this.addr && addr <= EndAddr;
-	public uint GetOffset(uint addr) => addr - this.addr;
+	public bool IsInRange(ushort addr) => addr >= this.addr && addr <= EndAddr;
+	public ushort GetOffset(ushort addr) => (ushort)(addr - this.addr);
 }
